split out reducer, create contexts
