+++ next-app/src/pages/index.tsx
@@ -1,47 +1,10 @@
-import { useEffect, useReducer } from 'react'
+import { useReducer, useEffect, createContext, Dispatch } from 'react'
 import PriceHistoryChart from '../components/chart'
-// import styles from  "../styles/UserHome.module.css"
-import { ItemProps, Item } from '../components/Item'
-// import { initialState, stateReducer } from '../utils/context'
-
-interface InitialState {
-  user: {
-    name: string,
-    id: number,
-    // favourites: ItemProps[],
-  },
-  items: null | ItemProps[],
-  activeItem: null | ItemProps
-}
-
-export const initialState: InitialState = {
-  user: {
-    name: 'Testy McTest face',
-    id: 1,
-    // favourites: []
-  },
-  items: null,
-  activeItem: null
-}
-
-
-export function stateReducer(state: typeof initialState, action: ACTIONTYPE) {
-  switch(action.type) {
-    case 'UPDATE-ITEMS':
-      if (!state.activeItem) {
-        return { ...state, items: action.payload, activeItem: action.payload[0]}
-      } else {
-        return {...state, items: action.payload}
-      }
-    default:
-      return state
-      throw new Error("bad action type")
-  }
-}
-
-type ACTIONTYPE = | { type: 'UPDATE-ITEMS', payload: ItemProps[] }
+import { ItemCard } from '../components/Item'
+import { initialState, stateReducer, StateContext, ACTIONTYPE} from '../lib/state-reducer'
 
 function UserHome(): JSX.Element {
+  
   const [state, dispatch] = useReducer(stateReducer, initialState)
 
   const getItems = useEffect(() => {
@@ -56,27 +19,32 @@ function UserHome(): JSX.Element {
     }
     getData()
   }, [])
+  const DispatchContext = createContext<Dispatch<ACTIONTYPE>>(dispatch)
 
-  console.log(state)
-
-  return (<div className='py-4 px-9 flex h-screen bg-slate-100'>
-    <main className='grid grid-cols-2'>
-      <div>
-        <h1 className='font-sans font-bold text-3xl py-3 m-2'>Items</h1>
-        <div className='grid grid-cols-1 gap-5'>
-          { state.items ? state.items.map(item => (
-            <Item   {...item}  active={item.id === state.activeItem?.id} key={item.id} />
-          ))
-          :
-          <></>
-          }
-        </div>
+  return (
+  <StateContext.Provider value={state}>
+    <DispatchContext.Provider value={dispatch}>
+      <div className='py-4 px-9 flex h-screen bg-slate-100'>
+        <main className='grid grid-cols-2'>
+          <div>
+            <h1 className='font-sans font-bold text-3xl py-3 m-2'>Items</h1>
+            <div className='grid grid-cols-1 gap-5'>
+              { state.items ? state.items.map(item => (
+                <ItemCard  {...item}  active={item.id === state.activeItem?.id} key={item.id}  />
+              ))
+              :
+              <></>
+              }
+            </div>
+          </div>
+          <div className='mt-11'>
+          {state.activeItem && <PriceHistoryChart activeItem={state.activeItem}/>}
+          </div>
+        </main>
       </div>
-      <div className='mt-11'>
-      {state.activeItem && <PriceHistoryChart activeItem={state.activeItem}/>}
-      </div>
-    </main>
-  </div>)
+    </DispatchContext.Provider>
+  </StateContext.Provider> 
+  )
 }
 
 export default UserHome
